A tidier version of chisq.test for goodness of fit tests and tests of independence.
Source: R/wrappers.R

A tidier version of chisq.test for goodness of fit tests and tests of independence.

chisq_test(data, formula, ...)

Arguments

data
a data frame that can be coerced into a tibble

formula
a formula with the response variable on the left and the explanatory on the right

...
additional arguments for chisq.test

Examples

# chisq test for comparing number of cylinders against automatic/manual
  mtcars %>%
    dplyr::mutate(cyl = factor(cyl), am = factor(am)) %>%
    chisq_test(cyl ~ am)

#> Warning: Chi-squared approximation may be incorrect

#>   statistic chisq_df    p_value
#> 1  8.740733        2 0.01264661