Deprecated functions and objects

These functions and objects should no longer be used. They will be removed in a future release of infer.

Usage

conf_int(x, level = 0.95, type = "percentile", point_estimate = NULL)

p_value(x, obs_stat, direction)

Arguments

x

See the non-deprecated function.

level

See the non-deprecated function.

type

See the non-deprecated function.

point_estimate

See the non-deprecated function.

obs_stat

See the non-deprecated function.

direction

See the non-deprecated function.

See also