Compute the confidence interval for (currently only) simulation-based methods

get_confidence_interval() and get_ci() are both aliases of conf_int()

conf_int(x, level = 0.95, type = "percentile", point_estimate = NULL)

get_ci(x, level = 0.95, type = "percentile", point_estimate = NULL)

get_confidence_interval(x, level = 0.95, type = "percentile",
  point_estimate = NULL)

Arguments

x
data frame of calculated statistics or containing attributes of theoretical distribution values. Currently, dependent on statistics being stored in stat column as created in calculate() function.

level
a numerical value between 0 and 1 giving the confidence level. Default value is 0.95.

type
a string giving which method should be used for creating the confidence interval. The default is "percentile" with "se" corresponding to (2 * standard error) as the other option.

point_estimate
a numeric value or a 1x1 data frame set to NULL by default. Needed to be provided if type = "se".

Value

a 2 x 1 tibble with values corresponding to lower and upper values in the confidence interval

Examples

mtcars_df <- mtcars %>%
    dplyr::mutate(am = factor(am))
d_hat <- mtcars_df %>%
    specify(mpg ~ am) %>%
    calculate(stat = "diff in means", order = c("1", "0"))
bootstrap_distn <- mtcars_df %>%
    specify(mpg ~ am) %>%
    generate(reps = 100) %>%
   calculate(stat = "diff in means", order = c("1", "0"))
bootstrap_distn %>% conf_int(level = 0.9)

#> # A tibble: 1 x 2
#>    `5%` `95%`
#>   <dbl> <dbl>
#> 1  4.27  10.1

bootstrap_distn %>% conf_int(type = "se", point_estimate = d_hat)

#> # A tibble: 1 x 2
#>   lower upper
#>   <dbl> <dbl>
#> 1  3.56  10.9