infer: a grammar for statistical inference

The objective of this package is to perform statistical inference using a grammar that illustrates the underlying concepts and a format that coheres with the tidyverse.

Details

For an overview of how to use the core functionality, see vignette("infer")

See also

Author

Maintainer: Simon Couch simon.couch@posit.co (ORCID)

Authors:

Andrew Bray abray@reed.edu

Chester Ismay chester.ismay@gmail.com (ORCID)

Evgeni Chasnovski evgeni.chasnovski@gmail.com (ORCID)

Ben Baumer ben.baumer@gmail.com (ORCID)

Mine Cetinkaya-Rundel mine@stat.duke.edu (ORCID)

Other contributors:

Ted Laderas tedladeras@gmail.com (ORCID) [contributor]

Nick Solomon nick.solomon@datacamp.com [contributor]

Johanna Hardin Jo.Hardin@pomona.edu [contributor]

Albert Y. Kim albert.ys.kim@gmail.com (ORCID) [contributor]

Neal Fultz nfultz@gmail.com [contributor]

Doug Friedman doug.nhp@gmail.com [contributor]

Richie Cotton richie@datacamp.com (ORCID) [contributor]

Brian Fannin captain@pirategrunt.com [contributor]