Specify the response and explanatory variables with specify also converting character variables chosen to be factors

Specify the response and explanatory variables with specify also converting character variables chosen to be factors

specify(x, formula, response = NULL, explanatory = NULL)

Arguments

x
a data frame that can be coerced into a tibble

formula
a formula with the response variable on the left and the explanatory on the right

response
the variable name in x that will serve as the response. This is alternative to using the formula argument.

explanatory
the variable name in x that will serve as the explanatory variable

Examples

# Response attribute set corresponding to
# response argument and specified variable selected
if(require(dplyr)){
  mtcars %>%
    mutate(am = factor(am)) %>%
    specify(response = am)
}

#> Response:  am 
#>  Response Type:  factor 
#> # A tibble: 32 x 1
#>        am
#>    <fctr>
#>  1      1
#>  2      1
#>  3      1
#>  4      0
#>  5      0
#>  6      0
#>  7      0
#>  8      0
#>  9      0
#> 10      0
#> # ... with 22 more rows

# Response and explanatory attributes set corresponding to
# response and explanatory arguments
if(require(dplyr)){
  mtcars %>%
    mutate(am = factor(am), vs = factor(vs)) %>%
    specify(response = am, explanatory = vs)
}

#> Response:  am 
#>  Response Type:  factor 
#> Explanatory:  vs 
#>  Explanatory Type:  factor 
#> # A tibble: 32 x 2
#>        am     vs
#>    <fctr> <fctr>
#>  1      1      0
#>  2      1      0
#>  3      1      1
#>  4      0      1
#>  5      0      0
#>  6      0      1
#>  7      0      0
#>  8      0      1
#>  9      0      1
#> 10      0      1
#> # ... with 22 more rows

# Response and explanatory attributes set corresponding to
# formula argument expecting response ~ explanatory
if(require(dplyr)){
  mtcars %>%
    mutate(am = factor(am), vs = factor(vs)) %>%
    specify(formula = am ~ vs)
}

#> Response:  am 
#>  Response Type:  factor 
#> Explanatory:  vs 
#>  Explanatory Type:  factor 
#> # A tibble: 32 x 2
#>        am     vs
#>    <fctr> <fctr>
#>  1      1      0
#>  2      1      0
#>  3      1      1
#>  4      0      1
#>  5      0      0
#>  6      0      1
#>  7      0      0
#>  8      0      1
#>  9      0      1
#> 10      0      1
#> # ... with 22 more rows