Specify the response and explanatory variables with specify also converting character variables chosen to be factors

Specify the response and explanatory variables with specify also converting character variables chosen to be factors

specify(x, formula, response = NULL, explanatory = NULL, success = NULL)

Arguments

x
a data frame that can be coerced into a tibble

formula
a formula with the response variable on the left and the explanatory on the right

response
the variable name in x that will serve as the response. This is alternative to using the formula argument

explanatory
the variable name in x that will serve as the explanatory variable

success
the level of response that will be considered a success, as a string. Needed for inference on one proportion, a difference in proportions, and corresponding z stats

Value

A tibble containing the response (and explanatory, if specified) variable data

Examples

# Permutation test similar to ANOVA
  mtcars %>%
    dplyr::mutate(cyl = factor(cyl)) %>%
    specify(mpg ~ cyl) %>%
    hypothesize(null = "independence") %>%
    generate(reps = 100, type = "permute") %>%
    calculate(stat = "F")

#> Response: mpg (numeric)
#> Explanatory: cyl (factor)
#> Null Hypothesis:  independence 
#> # A tibble: 100 x 2
#>    replicate   stat
#>        <int>  <dbl>
#>  1         1 0.0734
#>  2         2 0.0601
#>  3         3 0.177 
#>  4         4 0.665 
#>  5         5 0.163 
#>  6         6 1.88  
#>  7         7 0.663 
#>  8         8 0.148 
#>  9         9 1.82  
#> 10        10 0.250 
#> # ... with 90 more rows